Create clusterGroupUpgrade for each managedcluster
Create a new controller to watch managedCluster.
When a new managedcluster is created and ready, a ClusterGroupUpgrade
CR for this cluster will be created under namespace "ztp-install" with
the content of cluster name, list of policies and maxConcurrency.

The list of policies is generated from cluster's child policies and
the order is built based on the policy ztp-deploy-wave annotation.

Signed-off-by: Angie Wang <angwang@redhat.com>