Update sample manifests
The manifestsfrom samples are no longer valid, since the CRD schema
changed over the course of the past months.
I'm adding to this change a CGU CR and an upgrade policy, coupled
with a placement rule and placement binding for cluster called
'sno', that would do an upgrade to 4.9.4 version.
This samples can be used for testing an upgrade with the controller
without using ZTP at all.