Update Makefile to support PATH containing spaces
Prior to this update, the PATH variable is being modified and passed
on the SHELL command line. In an environment with spaces in the PATH,
such as WSL2, this causes the SHELL to be interpreted incorrectly and
causes all make commands to fail.

This update exports the modified PATH to make it available in the
environment, rather than passing it on the SHELL command line.

Signed-off-by: Don Penney <dpenney@redhat.com>