Add universal index processing
OLM catalogs are switching from sqlite-based to file-system based
index format (also referenced as FBC). The enhancement is described in
https://github.com/operator-framework/enhancements/blob/master/enhancements/declarative-index-config.md
Additional information is available hete:
https://github.com/redhat-openshift-ecosystem/community-operators-prod/discussions/512
The switch to FBC is expected to occur in Openshift 4.11.
This commit changes the way catalog processing is
done by pre-caching workload to extract
packages, bundles and channels relations from the
index. It is switching to a new opm command
(`opm render`), that works for both sqlite-based catalogs and FBC.

Signed-off-by: Vitaly Grinberg <vgrinber@redhat.com>