Add verify-bindata and shellcheck to ci-job
This update adds the verify-bindata and shellcheck tests to ci-job in
the Makefile, which is run as part of the CI integration test. This
will protect against the introduction of changes to bindata that has
not been reflected in the generated golang source, as well as new
shellcheck warnings in bash scripts.

Signed-off-by: Don Penney <dpenney@redhat.com>