Fix OLM index extraction
This fixes a bug in sqlite OLM catalog extraction, causing
operator bundle to be identified incorrectly. The bug was
percieved as a failure to pull one or more images.
The problem is fixed by cherry-picking the following commit from the
main branch:

Commit d019df2
This fixes an authorization problem occurring when
trying to render an OLM index container with opm.
There are two tools used in the precaching workload
from inside the container, podman and opm.
Podman is using auth.json from /run/containers/0
when run as root. Opm is using ~/.docker/config.json.
This change adds missing pull secret copy to the
latter for the opm use, and also puts all the relevant
copy operations to the "copy_env.sh" script

Signed-off-by: Vitaly Grinberg <vgrinber@redhat.com>