Use builtin client in createResourcesFromTemplates
Unit tests relay on the controller fake client. Using other clients make
code impossible to test.
This commit removes one of the instances where dynamic client was
used making it possible to unit-tests it.
In addition, it fixes a bug discovered in progress, where template
was rendered with uninitialized data.