Add precaching pull spec as env var
This is a manual cherrypick of pull/155. It is fixing a
productization issue with digest pinning, when the
related images field in the csv is not patched if it
is not empty.
The fix propagates the pre-caching image pull spec as
the operator deployment environment variable.
It is created when running `make bundle`, and should
be patched by the productization machinery after building the
pre-caching image with the correct sha256 digest.

Signed-off-by: Vitaly Grinberg <vgrinber@redhat.com>