Make image tags consistent with upstream builds
This makes image tags of the deployment and related images consistent
with the tags applied in quay to the images pushed by the upstream CI.
This is needed for installing the upstream built operator.
If this is not done, the operator installation will fail because of
image pull error.
In addition it reverts the operator-sdk version to the latest
one recommended by the OCP documentation (operator-sdk-v1.16.0-ocp).