recovery: Restore /var/lib/kubelet after reboot
Moved the restore of /var/lib/kubelet from the first stage of recovery
to the start of the second stage, after the reboot. As kubelet.service
is shutdown and disabled during the first stage to prevent it coming
up after the reboot until launched by the recovery utility, this will
ensure there are no open references to these files that would cause
the restore rsync to fail.

Signed-off-by: Don Penney <dpenney@redhat.com>