Bug 2075585: Support for hub cluster template function fromConfigMap
ACM hub cluster resource-specific template functions are restricted to the same
namespace as the policy is defined on the hub cluster due to security concerns,
so TALM cannot work with hub cluster template as TALM creates the copied policies
in the namespace where the CGU is created. This commit updates to support the
hub-side resource-specific function fromConfigMap by replicating the configmap
to the CGU namespace.

Changes included:
 - validate managed policies before reconciling resources, requeue with medium interval
   and update CGU status when invalid policies found
 - search for hub templates in each managed policy and replicate configmap in the CGU namespace
 - update the policy hub template with replicated configmap
 - requeue with medium interval when unsupported templates found
   i.e. printf variable in resource Name/Namespace field
        fromSecret/lookup functions
 - unittest for valid and invalid hub templates usage