Use overlayfs storage driver (#330)
This commit is fixing the following error in pre-caching:
'libsubid.so.3: cannot open shared object file:
 No such file or directory\n"'
The libsubid.so.3 is a podman dependency used when
container storage is configured with vfs driver.
This library was absent from UBI images for some time
(https://bugzilla.redhat.com/show_bug.cgi?id=2084179)
Pre-caching implementation had a remainder from experimenting
with vfs driver, that this commit removes, eliminating the
need for libsubid.so.3.