Add ability to wait for `soak duration` after policy becomes compliant
If a policy has the `soakSeconds` annotation, the controller will wait for the specified
number of seconds after the first time it observes that the cluster is compliant with
that policy before moving on from that policy. The controller will set
`status.status.currentBatchRemediationProgress.clusterName.firstCompliantAt` the first time
it observes that the cluster is compliant with the policy. In the following reconciliations,
if the requested number of seconds has passed since firstCompliantAt, and the cluster remains
compliant with the policy, the controller will move on from that policy.

Signed-off-by: Saeid Askari <saskari@redhat.com>