Remove target directory in backup function of upgrade-recovery script (#659)
On a failed copy attempt, a retry will result in the source directory being copied under a partially copied target directory.
This results in a nested directory structure in which the nested directory holds the most recent copied contents.

This commit removes the target directory specification from the copy command cp -Ra.

Signed-off-by: Sharat Akhoury <sakhoury@redhat.com>
Co-authored-by: Sharat Akhoury <sakhoury@redhat.com>