CNF-11189: Add AZTP binary in a separate image
Monitor ManagedCluster resource for AZTP label
If labelled, extract child policies for the cluster and wrap
the resources in the AZTP configmap.
If labelled for full acceleration, add manifests
for AZTP post-reboot service.
This commit also includes AZTP post reboot service source code
under pkg/aztp. It is built and copied into a separate image
for post-reboot execution on spokes

Signed-off-by: Vitaly Grinberg <v.g@redhat.com>