CNF-13181: Allow partial completion before moving to next CGU
Before this when multiple CGUs are chain together, the first CGU must
completely finish before the next CGU is allowed to start. This means
that if one cluster fails the next CGU can not be started for the rest
of clusters.

Now by setting the blockingCRCompletionMode annotation to `partial`, the dependent CGU starts
for the clusters that are compliant with the previous CGUs.

Signed-off-by: Saeid Askari <saskari@redhat.com>

Update controllers/clustergroupupgrade_controller.go

Co-authored-by: Angie Wang <angwang@redhat.com>
Signed-off-by: Saeid Askari <saskari@redhat.com>