Add `ran.openshift.io/blocking-cgu-cluster-filtering` annotation. (#800)
This commit introduces the `ran.openshift.io/blocking-cgu-cluster-filtering`
annotation. When `ran.openshift.io/blocking-cgu-completion-mode` is set
to `partial` and `ran.openshift.io/blocking-cgu-cluster-filtering` is set
to `true`, the behavior is modified such that if a (CGU) fails, the subsequent
 CGU continues only on the clusters where the blocking CGU was successful.

If `ran.openshift.io/blocking-cgu-cluster-filtering` is not set to `true`,
the next CGU continues on all clusters, regardless of their success status.

Signed-off-by: Saeid Askari <saskari@redhat.com>