Add `Idle` action for IBGU
This commit removes the single actions 'Finalize' and 'Abort' from the
valid set due to ambiguity in their outcomes depending on the cluster's
current state. Instead, it introduces an 'Idle' action. Unlike
'Finalize' or 'Abort', 'Idle' does not specify whether it will result in
finalization or aborting of actions. The outcome will be determined by
the cluster's state at the time of execution, requiring users to be
aware of this behavior.

Signed-off-by: Saeid Askari <saskari@redhat.com>